+++ django_countries/tests/graphql/test_country_type.py
@@ -1,0 +1,17 @@
+from graphene.test import Client
+from django_countries.tests.graphql.schema import schema
+
+
+def test_country_type():
+    client = Client(schema)
+    executed = client.execute(
+        """{ newZealand {code, name, iocCode, numeric, alpha3} }"""
+    )
+    returned_country = executed["data"]["newZealand"]
+    assert returned_country == {
+        "code": "NZ",
+        "name": "New Zealand",
+        "iocCode": "NZL",
+        "numeric": 554,
+        "alpha3": "NZL",
+    }